requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
openai>=1.30.0
beautifulsoup4==4.12.2
lxml>=4.9.0
playwright>=1.40.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0